requirements.txt
# ============================================================================
# Deep Learning & Computer Vision
# ============================================================================
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# ============================================================================
# Tracking Libraries
# ============================================================================
boxmot>=10.0.0
filterpy>=1.4.5
lap>=0.4.0

# ============================================================================
# MOT Evaluation
# ============================================================================
motmetrics>=1.4.0

# ============================================================================
# Data Processing
# ============================================================================
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ============================================================================
# Visualization
# ============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0

# ============================================================================
# Utilities
# ============================================================================
tqdm>=4.65.0
PyYAML>=6.0
configparser>=5.3.0

# ============================================================================
# Optional: For Jupyter Notebook
# ============================================================================
# jupyter>=1.0.0
# notebook>=7.0.0
# ipywidgets>=8.0.0
